+++ backend/config/routes.php
@@ -54,6 +54,8 @@ return [
     '/api/app/v1/esi/eve-login/{name}/characters'   => ['GET',    [AppEsiController::class, 'eveLoginCharacters']],
     '/api/app/v1/esi[{path:.*}]'                    => ['GET'  => [AppEsiController::class, 'esiV1'],
                                                         'POST' => [AppEsiController::class, 'esiPostV1']],
+    '/api/app/v2/esi[{path:.*}]'                    => ['GET'  => [AppEsiController::class, 'esiV2'],
+                                                        'POST' => [AppEsiController::class, 'esiPostV2']],
 
     '/api/app/v1/groups/{cid}'              => ['GET',  [AppGroupController::class, 'groupsV1']],
     '/api/app/v2/groups/{cid}'              => ['GET',  [AppGroupController::class, 'groupsV2']],