+++ backend/config/security.php
@@ -27,6 +27,7 @@ return [
     '/api/app/v1/corporation/{id}/member-tracking' => [Role::APP_TRACKING],
 
     '/api/app/v1/esi'                              => [Role::APP_ESI],
+    '/api/app/v2/esi'                              => [Role::APP_ESI],
 
     '/api/app/v1/groups'                           => [Role::APP_GROUPS],
     '/api/app/v2/groups'                           => [Role::APP_GROUPS],