+++ CHANGELOG.md
@@ -6,6 +6,9 @@ dd mm yyyy
 
 - **BC Break**: The `/core.alt` and `/core.managed-alt` login URLs have been removed. Use `core.default` and 
   `core.managed` respectively.
+- Fixed type of Expires header from integer to string in OpenAPI definition file for /app/v1/esi.
+- Added /app/v2/esi and deprecated /app/v1/esi. The new version contains error messages in the body
+  instead of the reason phrase.
 
 ## 1.32.0